# api/openapi.yaml
openapi: 3.1.0
info:
  title: leani native API
  version: 1.0.0
  description: |
    Coverage-aware queries and durable processor changes. All cursors are
    opaque. Ethereum quantities that can exceed JavaScript's safe integer
    range are encoded as unsigned decimal strings.

    Every route answers a `Host` name other than `localhost` and those in
    `api.allowed_hosts` with 421 `host_not_allowed` (IP addresses pass), and
    a browser `Origin` other than a loopback one or one in
    `api.allowed_origins` with 403 `origin_not_allowed`. A GET or HEAD
    without an `Origin` whose `Sec-Fetch-Site` is `cross-site` or
    `same-site` and whose `Sec-Fetch-Mode` is not `navigate` gets 403
    `cross_site_request`; on the two consumer stream routes, so does one in
    any mode, navigations included. A POST or DELETE needs
    `content-type: application/json` or `x-leani-request: 1`, and gets 415
    `unsupported_media_type` otherwise.

    A consumer created with a credential needs it in
    `x-leani-consumer-credential` on every consumer-scoped delivery route,
    in addition to any bearer token, and gets 403 `forbidden` without it.
    Listing, reading, and revoking consumers are administrative: only the
    bearer token protects them, whether or not a consumer has a credential.
servers:
  - url: http://127.0.0.1:8080
security:
  - {}
  - bearerAuth: []
paths:
...
  /v1/backfill-subscriptions/{subscription}/consumers/{consumer}/stream:
    parameters:
      - $ref: "#/components/parameters/Subscription"
      - $ref: "#/components/parameters/Consumer"
      - $ref: "#/components/parameters/ConsumerCredential"
    get:
      summary: Stream changes from one backfill subscription
      operationId: streamBackfillConsumer
      description: |
        Each batch and the completion record the consumer's delivered
        sequence through their acknowledgement boundary. A request whose
        `Sec-Fetch-Site` is `cross-site` or `same-site` without an allowed
        `Origin`, such as a navigation of another site's frame, gets 403
        `cross_site_request`. When the node shuts down, the stream ends after
        a whole record; reconnect to resume after the acknowledged cursor.
      responses:
        "200":
          description: Server-sent backfill delivery batches.
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    Consumer:
      name: consumer
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    ConsumerCredential:
      name: x-leani-consumer-credential
      in: header
      required: false
      description: |
        The consumer's credential. A consumer created with one needs it, in
        addition to any bearer token, on its lease, change, acknowledgement,
        and stream requests, and gets 403 `forbidden` without it or with
        another; the node compares it in constant time. Listing, inspecting,
        and revoking consumers need only the bearer token. A consumer without
        one, such as a configured consumer, takes none. Credentials created
        before 32 characters were required keep working.
      schema:
        type: string
        minLength: 16
        maxLength: 512
...
    Subscription:
      name: subscription
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    Error:
      description: Stable machine-readable error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
...
    ErrorBody:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            details:
              oneOf:
                - type: object
                  additionalProperties: true
                - type: "null"
            requestId:
              type: string